changed a bunch: - made UI better (i hope) - set gps base and set pressure base messages sent.  that still needs work though. - button added to turn on and off FC. - abort button now actually aborts FC (sequencer) along with countdown.
todo still:

clean up and review.  code is ugly.  it is ugly because i'm not sure
the best way to do gui stuff, so lots of it is commented out.
it is also ugly from lots of developement and needs more review.

RocketState needs to have text of current state, and also grid containing
info about state detail.


git-svn-id: http://svn.psas.pdx.edu/svn/psas/trunk/java@1341 6115d4fc-9e01-0410-8836-dad5f2c52705